Added more func to courtController

+++ TennisApp/src/main/java/cz/inqool/tennisapp/domain/court/CourtController.java
@@ -4,6 +4,7 @@ import cz.inqool.tennisapp.domain.reservation.Reservation;
 import cz.inqool.tennisapp.domain.reservation.ReservationService;
 import cz.inqool.tennisapp.utils.exceptions.NotFoundException;
 import cz.inqool.tennisapp.utils.response.ArrayResponse;
+import cz.inqool.tennisapp.utils.response.ObjectResponse;
 import jakarta.transaction.Transactional;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -30,7 +31,14 @@ public class CourtController {
         );
     }
 
-    
+    @GetMapping(value = "/{id}", produces = "application/json")
+    public ObjectResponse<CourtResponse> getCourtById(@PathVariable int id) {
+        Court court = courtService.getCourtById(id);
+        if (court == null) {
+            throw new NotFoundException();
+        }
+        return ObjectResponse.of(court, CourtResponse::new);
+    }
 
 
     @GetMapping("/{courtId}/reservations")
@@ -45,8 +53,10 @@ public class CourtController {
     @Transactional
     public void deleteCourt(@PathVariable int id) {
         Court court = courtService
-                .getCourtById(id)
-                .orElseThrow(NotFoundException::new);
+                .getCourtById(id);
+        if (court == null) {
+            throw new NotFoundException();
+        }
         // only SOFT Delete
         courtService.deleteCourt(court);
     }
